selvbetjening/requirements.txt
django==2.2.18
gunicorn==20.0.4
psycopg2==2.8.6
fpdf==1.7.2
django-cookies-samesite>=0.8.0
pyjwkest>=1.3.6
oic>=1.2.0
zeep>=3.4.0
